Correctly handle the last bit for the stone plane hex strings
For a board with an odd number of rows and columns,
there are a total of (2*k+1)^2 = 4*k^2 + 4*k + 1 positions or "bits"
If we write this out as a hexadecimal, i.e. group by 4 bits (2^4 = 16),
there will always be 1 bit left, i.e. (4*k^2 + 4*k + 1) % 4 = 1.

LeelaZero handles this in a unique way:
Instead of treating the bit array as one hexadecimal (by prepending it with 0s to make the length divisible by 4),
it just appends the last bit at the end as 0 or 1. So we first need to parse the hex string without the last
digit, then append a bit to the parsed bit array at the end.

More details in the code below:
https://github.com/leela-zero/leela-zero/blob/b259e50d5cce34a12176846534f369ef5ffcebc1/src/Training.cpp\#L260-L264